+++ lib/solr_cloud/collection.rb
@@ -1,12 +1,42 @@
 # frozen_string_literal: true
 
 require "solr_cloud/connection"
+require "forwardable"
 
 module SolrCloud
   # A Collection provides basic services on the collection -- checking its health,
   # creating or reporting aliases, and deleting itself.
   class Collection
-    attr_reader :name, :connection
+
+    extend Forwardable
+
+    # Forward the HTTP verbs to the  underlying connection
+
+    # @!method get
+    # Forwarded on to the underlying SolrCloud connection
+    # @see SolrCloud::Connection#get
+    def_delegator :@connection, :get
+
+    # @!method post
+    # Forwarded on to the underlying SolrCloud connection
+    # @see SolrCloud::Connection#post
+    def_delegator :@connection, :post
+
+    # @!method delete
+    # Forwarded on to the underlying SolrCloud connection
+    # @see SolrCloud::Connection#delete
+    def_delegator :@connection, :delete
+
+    # @!method put
+    # Forwarded on to the underlying SolrCloud connection
+    # @see SolrCloud::Connection#put
+    def_delegator :@connection, :put
+
+    # @return [SolrCloud::Connection] the underlying SolrCloud connection
+    attr_reader :name
+
+    # @return The name of the collection
+    attr_reader :connection
 
     # In general, users shouldn't use Collection.new; instead, use
     # connection.create_collection(name: "coll_name", configset: "existing_configset_name")
@@ -104,6 +134,29 @@ module SolrCloud
     # @return [SolrCloud::ConfigSet]
     def configset
       Configset.new(name: info["configName"], connection: connection)
+    end
+
+    # Add a document or array of documents
+    # @todo Gotta check for errors and such!
+    # @param docs [Hash, Array<Hash>] One or more documents to add
+    # @return self
+
+    def add(docs)
+      docarray = if docs === Array
+        docs
+      else
+        [docs]
+      end
+      post("solr/#{name}/update/") do |r|
+        r.body = docarray
+      end
+      self
+    end
+
+    # Get a count of how many documents are in the index
+    # @return [Fixnum] the number of documents
+    def count
+      get("solr/#{name}/select", q: "*:*").body["response"]["numFound"]
     end
 
     def inspect