+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## [Unreleased]
+# Changelog
+
+## [0.2.0] - 2023-12-01
+
+- Added options `:date` and `:datetime` to the `version:` argument to `create_collection`
+  to automatically generate, e.g., "2023-12-01" or "2023-12-01-09-50-56" 
+- Added utility method `legal_solr_name?` to check for validity for collection names
 
 ## [0.1.0] - 2023-11-29